fix(ui): unified public key view screens
