+++ client/react-native/common/components/Library/index.js
@@ -13,3 +13,4 @@ export CustomTextInput from './CustomTextInput.js'
 export Badge from './Badge'
 export ModalScreen from './ModalScreen'
 export SearchBar from './SearchBar'
+export PublicKeyWithActions from './PublicKeyWithActions'
