+++ client/react-native/common/components/Screens/Contacts/Detail/PublicKey.js
@@ -1,19 +1,12 @@
 import React, { PureComponent } from 'react'
-import { Clipboard } from 'react-native'
 import { colors } from '../../../../constants'
 import {
-  Flex,
   Screen,
   Header,
-  Button,
-  TextInputMultilineFix,
+  PublicKeyWithActions,
 } from '../../../Library'
 import {
-  padding,
   paddingVertical,
-  textTiny,
-  marginTop,
-  rounded,
 } from '../../../../styles'
 
 export default class DetailPublicKey extends PureComponent {
@@ -23,47 +16,13 @@ export default class DetailPublicKey extends PureComponent {
   })
 
   render () {
+    console.log(this.props)
     const id = this.props.navigation.getParam('id')
+    const displayName = this.props.navigation.getParam('displayName')
+
     return (
       <Screen style={[{ backgroundColor: colors.white }, paddingVertical]}>
-        <Flex.Rows style={[padding]} align='center'>
-          <TextInputMultilineFix
-            style={[
-              {
-                width: 330,
-                height: 330,
-                backgroundColor: colors.grey7,
-                color: colors.black,
-                flexWrap: 'wrap',
-              },
-              textTiny,
-              padding,
-              marginTop,
-              rounded,
-            ]}
-            multiline
-            value={id}
-            selectTextOnFocus
-          />
-          <Flex.Cols align='start'>
-            <Flex.Rows>
-              <Button
-                icon='copy'
-                background={colors.blue}
-                margin
-                padding
-                rounded={23}
-                height={24}
-                medium
-                middle
-                center
-                onPress={() => Clipboard.setString(id)}
-              >
-                COPY THE KEY
-              </Button>
-            </Flex.Rows>
-          </Flex.Cols>
-        </Flex.Rows>
+        <PublicKeyWithActions initialKey={id} initialName={displayName} copyButton shareButton readOnly self={false} />
       </Screen>
     )
   }
