+++ client/react-native/common/components/Screens/Settings/MyAccount/MyPublicKey.js
@@ -1,21 +1,14 @@
 import React, { PureComponent } from 'react'
-import { Clipboard } from 'react-native'
 import { colors } from '../../../../constants'
 import {
-  Flex,
   Screen,
   Header,
-  Button,
-  TextInputMultilineFix,
+  PublicKeyWithActions,
 } from '../../../Library'
 import {
-  padding,
   paddingVertical,
-  textTiny,
-  marginTop,
-  rounded,
 } from '../../../../styles'
-import { atob } from 'b64-lite'
+import { extractPublicKeyFromId } from '../../../../helpers/contacts'
 
 export default class MyPublicKey extends PureComponent {
   static navigationOptions = ({ navigation }) => ({
@@ -23,69 +16,12 @@ export default class MyPublicKey extends PureComponent {
     header: <Header navigation={navigation} title='My public key' backBtn />,
   })
 
-  share = () => {
-    const { id } = this.props.navigation.getParam('data')
-    console.log('Share: ', id)
-  }
-
   render () {
-    const { id } = this.props.navigation.getParam('data')
-    const myID = atob(id).split('contact:')[1]
+    const { id, displayName } = this.props.navigation.getParam('data')
+    const myID = extractPublicKeyFromId(id)
     return (
       <Screen style={[{ backgroundColor: colors.white }, paddingVertical]}>
-        <Flex.Rows style={[padding]} align='center'>
-          <TextInputMultilineFix
-            style={[
-              {
-                width: 330,
-                height: 330,
-                backgroundColor: colors.grey7,
-                color: colors.black,
-                flexWrap: 'wrap',
-              },
-              textTiny,
-              padding,
-              marginTop,
-              rounded,
-            ]}
-            multiline
-            placeholder='Type or copy/paste a berty user public key here'
-            value={myID}
-            selectTextOnFocus
-          />
-          <Flex.Cols align='start'>
-            <Flex.Rows>
-              <Button
-                icon={'share'}
-                background={colors.blue}
-                margin
-                padding
-                rounded={23}
-                height={24}
-                medium
-                middle
-                center
-                onPress={this.share}
-              >
-                SHARE THE KEY
-              </Button>
-              <Button
-                icon='copy'
-                background={colors.blue}
-                margin
-                padding
-                rounded={23}
-                height={24}
-                medium
-                middle
-                center
-                onPress={() => Clipboard.setString(myID)}
-              >
-                COPY THE KEY
-              </Button>
-            </Flex.Rows>
-          </Flex.Cols>
-        </Flex.Rows>
+        <PublicKeyWithActions initialKey={myID} initialName={displayName} copyButton shareButton readOnly self />
       </Screen>
     )
   }
